requirements.txt
pandas==1.5.1
matplotlib
Flask==2.2.2
Werkzeug==2.2.3
pydantic==1.10.21
seaborn
pytest==7.2.2
requests==2.28.2
numpy==1.24.3
fastapi==0.115.12
openai==1.76.2
python-dotenv==1.1.0
scikit_learn==1.6.1
uvicorn==0.34.2
rdflib==7.0.0
